+++ css/landing_page.css
@@ -9,7 +9,77 @@
     height: auto;
 }
 
+/* Arrow */
 
+@-moz-keyframes bounce {
+    0%, 20%, 50%, 80%, 100% {
+      -moz-transform: translateY(0);
+      transform: translateY(0);
+    }
+    40% {
+      -moz-transform: translateY(-30px);
+      transform: translateY(-30px);
+    }
+    60% {
+      -moz-transform: translateY(-15px);
+      transform: translateY(-15px);
+    }
+  }
+  @-webkit-keyframes bounce {
+    0%, 20%, 50%, 80%, 100% {
+      -webkit-transform: translateY(0);
+      transform: translateY(0);
+    }
+    40% {
+      -webkit-transform: translateY(-30px);
+      transform: translateY(-30px);
+    }
+    60% {
+      -webkit-transform: translateY(-15px);
+      transform: translateY(-15px);
+    }
+  }
+  @keyframes bounce {
+    0%, 20%, 50%, 80%, 100% {
+      -moz-transform: translateY(0);
+      -ms-transform: translateY(0);
+      -webkit-transform: translateY(0);
+      transform: translateY(0);
+    }
+    40% {
+      -moz-transform: translateY(-30px);
+      -ms-transform: translateY(-30px);
+      -webkit-transform: translateY(-30px);
+      transform: translateY(-30px);
+    }
+    60% {
+      -moz-transform: translateY(-15px);
+      -ms-transform: translateY(-15px);
+      -webkit-transform: translateY(-15px);
+      transform: translateY(-15px);
+    }
+  }
+  
+  .arrow {
+    position: fixed;
+    bottom: 50px;
+    left: 50%;
+    margin-left: -20px;
+    width: 40px;
+    height: 40px;
+    background-image: url(../assets/arrow-down.png);
+    background-size: contain;
+  }
+  
+  .bounce {
+    -moz-animation: bounce 2s 3;
+    -webkit-animation: bounce 2s 3;
+    animation: bounce 2s 3;
+  }
+  .fade {
+    opacity:0;
+    transition: 1s all ease-out;
+  }
 /* Carousel */
 
 .landing-img-1 {
@@ -41,121 +111,17 @@
 }
 
 /* Ken Burns Effect */
-
-.carousel-image {
-    animation-timing-function: linear;
-    animation-iteration-count: infinite;
-    animation-duration: 16s;
-    transform: scale(1.2);
-}
-
-.carousel-image:nth-child(1) {
-    animation-name: kenburns-1;
-    z-index: 5;
-}
-.carousel-image:nth-child(2) {
-    animation-name: kenburns-2;
-    z-index: 4;
-}
-.carousel-image:nth-child(3) {
-    animation-name: kenburns-3;
-    z-index: 3;
-}
-.carousel-image:nth-child(4) {
-    animation-name: kenburns-4;
-    z-index: 2;
-}
-.carousel-image:nth-child(5) {
-    animation-name: kenburns-5;
-    z-index: 1;
-}
-
-@keyframes kenburns-1 {
+@keyframes kenBurns {
     0% {
-    	opacity: 0;
-    }
-    5% {
-    	opacity: 1;
-    }
-    95% {
-    	transform: scale3d(1.2, 1.2, 1.2) translate3d(69px, -2px, 0px);
-    	animation-timing-function: ease-in;
-    	opacity: 1;
+      transform: scale(1);
+      transform-origin: center;
     }
     100% {
-    	transform: scale3d(1.7, 1.7, 1.7) translate3d(-11px, -12px, 0px);
-    	opacity: 0;
+      transform: scale(1.1);
+      transform-origin: center;
     }
-}
-
-@keyframes kenburns-2 {
-    0% {
-    	opacity: 0;
-    }
-    5% {
-    	opacity: 1;
-    }
-    95% {
-    	transform: scale3d(1.2, 1.2, 1.2) translate3d(69px, -2px, 0px);
-    	animation-timing-function: ease-in;
-    	opacity: 1;
-    }
-    100% {
-    	transform: scale3d(1.7, 1.7, 1.7) translate3d(-11px, -12px, 0px);
-    	opacity: 0;
-    }
-}
-
-@keyframes kenburns-3 {
-    0% {
-    	opacity: 0;
-    }
-    5% {
-    	opacity: 1;
-    }
-    95% {
-    	transform: scale3d(1.2, 1.2, 1.2) translate3d(69px, -2px, 0px);
-    	animation-timing-function: ease-in;
-    	opacity: 1;
-    }
-    100% {
-    	transform: scale3d(1.7, 1.7, 1.7) translate3d(-11px, -12px, 0px);
-    	opacity: 0;
-    }
-}
-
-@keyframes kenburns-4 {
-    0% {
-    	opacity: 0;
-    }
-    5% {
-    	opacity: 1;
-    }
-    95% {
-    	transform: scale3d(1.2, 1.2, 1.2) translate3d(69px, -2px, 0px);
-    	animation-timing-function: ease-in;
-    	opacity: 1;
-    }
-    100% {
-    	transform: scale3d(1.7, 1.7, 1.7) translate3d(-11px, -12px, 0px);
-    	opacity: 0;
-    }
-}
-
-@keyframes kenburns-5 {
-    0% {
-    	opacity: 0;
-    }
-    5% {
-    	opacity: 1;
-    }
-    95% {
-    	transform: scale3d(1.2, 1.2, 1.2) translate3d(69px, -2px, 0px);
-    	animation-timing-function: ease-in;
-    	opacity: 1;
-    }
-    100% {
-    	transform: scale3d(1.7, 1.7, 1.7) translate3d(-11px, -12px, 0px);
-    	opacity: 0;
-    }
-}
+  }
+  
+  .carousel-image {
+    animation: kenBurns 10s infinite alternate; /* Adjust animation duration and other properties as needed */
+  }
