+++ css/portfolio.css
@@ -1,0 +1,37 @@
+.custom-border {
+    border: 5px 2px solid black;
+}
+.overlay-portfolio {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    color: white;
+    display: flex;
+    flex-direction: column;
+    justify-content: end;
+    /* align-items: center; */
+    /* position: absolute;
+    top: 80%;
+    left: 20%;
+    transform: translate(-50%, -50%);
+    text-align: center;
+    color: white; */
+}
+.overlay-portfolio:hover {
+    background: linear-gradient(to bottom, rgba(34, 34, 34, 0.6) 0%, rgba(34, 34, 34, 0.5) 100%);
+}
+
+.overlay-portfolio > a {
+  display: none;
+  margin-left: 10px;
+  margin-bottom: 10px;
+}
+.overlay-portfolio:hover > a {
+  display: block;
+}
+
+.overlay-portfolio > a:hover {
+  /* cursor: pointer; */
+}
